test [nfc]: Consolidate a widgets/checks.dart for trivial checks-extensions
I've long had a recurring small annoyance that the "foo_checks.dart"
file interferes with tab-completion for "foo_test.dart".  For example
as mentioned in this comment:
  https://github.com/zulip/zulip-flutter/pull/1608#discussion_r2155918312

So here's an approach to solve that, beginning with the "checks" files
in test/widgets/.

For updating the imports, I used a bit of Perl:

  $ perl -i -0pe '
        s,import .\K(?!dialog)\w*_checks.dart,checks.dart,g
      ' test/widgets/*_test.dart
