+++ test/widgets/action_sheet_test.dart
@@ -44,7 +44,7 @@ import '../stdlib_checks.dart';
 import '../test_clipboard.dart';
 import '../test_images.dart';
 import '../test_share_plus.dart';
-import 'compose_box_checks.dart';
+import 'checks.dart';
 import 'dialog_checks.dart';
 import 'test_app.dart';
