+++ test/widgets/channel_colors_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter_checks/flutter_checks.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:zulip/widgets/channel_colors.dart';
 
-import 'channel_colors_checks.dart';
+import 'checks.dart';
 
 void main() {
   group('ChannelColorSwatches', () {
