+++ test/widgets/checks.dart
@@ -1,0 +1,86 @@
+import 'package:checks/checks.dart';
+import 'package:flutter/widgets.dart';
+
+import 'package:zulip/model/narrow.dart';
+import 'package:zulip/widgets/channel_colors.dart';
+import 'package:zulip/widgets/compose_box.dart';
+import 'package:zulip/widgets/content.dart';
+import 'package:zulip/widgets/message_list.dart';
+import 'package:zulip/widgets/page.dart';
+import 'package:zulip/widgets/profile.dart';
+import 'package:zulip/widgets/store.dart';
+import 'package:zulip/widgets/unread_count_badge.dart';
+
+extension ChannelColorSwatchChecks on Subject<ChannelColorSwatch> {
+  Subject<Color> get base => has((s) => s.base, 'base');
+  Subject<Color> get unreadCountBadgeBackground => has((s) => s.unreadCountBadgeBackground, 'unreadCountBadgeBackground');
+  Subject<Color> get iconOnPlainBackground => has((s) => s.iconOnPlainBackground, 'iconOnPlainBackground');
+  Subject<Color> get iconOnBarBackground => has((s) => s.iconOnBarBackground, 'iconOnBarBackground');
+  Subject<Color> get barBackground => has((s) => s.barBackground, 'barBackground');
+}
+
+extension ComposeBoxStateChecks on Subject<ComposeBoxState> {
+  Subject<ComposeBoxController> get controller => has((c) => c.controller, 'controller');
+}
+
+extension ComposeBoxControllerChecks on Subject<ComposeBoxController> {
+  Subject<ComposeContentController> get content => has((c) => c.content, 'content');
+  Subject<FocusNode> get contentFocusNode => has((c) => c.contentFocusNode, 'contentFocusNode');
+}
+
+extension StreamComposeBoxControllerChecks on Subject<StreamComposeBoxController> {
+  Subject<ComposeTopicController> get topic => has((c) => c.topic, 'topic');
+  Subject<FocusNode> get topicFocusNode => has((c) => c.topicFocusNode, 'topicFocusNode');
+}
+
+extension EditMessageComposeBoxControllerChecks on Subject<EditMessageComposeBoxController> {
+  Subject<int> get messageId => has((c) => c.messageId, 'messageId');
+  Subject<String?> get originalRawContent => has((c) => c.originalRawContent, 'originalRawContent');
+}
+
+extension ComposeContentControllerChecks on Subject<ComposeContentController> {
+  Subject<List<ContentValidationError>> get validationErrors => has((c) => c.validationErrors, 'validationErrors');
+}
+
+extension RealmContentNetworkImageChecks on Subject<RealmContentNetworkImage> {
+  Subject<Uri> get src => has((i) => i.src, 'src');
+  // TODO others
+}
+
+extension AvatarImageChecks on Subject<AvatarImage> {
+  Subject<int> get userId => has((i) => i.userId, 'userId');
+}
+
+extension AvatarShapeChecks on Subject<AvatarShape> {
+  Subject<double> get size => has((i) => i.size, 'size');
+  Subject<double> get borderRadius => has((i) => i.borderRadius, 'borderRadius');
+  Subject<Widget> get child => has((i) => i.child, 'child');
+}
+
+extension MessageListPageChecks on Subject<MessageListPage> {
+  Subject<Narrow> get initNarrow => has((x) => x.initNarrow, 'initNarrow');
+  Subject<int?> get initAnchorMessageId => has((x) => x.initAnchorMessageId, 'initAnchorMessageId');
+}
+
+extension WidgetRouteChecks<T> on Subject<WidgetRoute<T>> {
+  Subject<Widget> get page => has((x) => x.page, 'page');
+}
+
+extension AccountRouteChecks<T> on Subject<AccountRoute<T>> {
+  Subject<int> get accountId => has((x) => x.accountId, 'accountId');
+}
+
+extension ProfilePageChecks on Subject<ProfilePage> {
+  Subject<int> get userId => has((x) => x.userId, 'userId');
+}
+
+extension PerAccountStoreWidgetChecks on Subject<PerAccountStoreWidget> {
+  Subject<int> get accountId => has((x) => x.accountId, 'accountId');
+  Subject<Widget> get child => has((x) => x.child, 'child');
+}
+
+extension UnreadCountBadgeChecks on Subject<UnreadCountBadge> {
+  Subject<int> get count => has((b) => b.count, 'count');
+  Subject<bool> get bold => has((b) => b.bold, 'bold');
+  Subject<Color?> get backgroundColor => has((b) => b.backgroundColor, 'backgroundColor');
+}
