+++ test/widgets/home_test.dart
@@ -24,8 +24,7 @@ import '../model/binding.dart';
 import '../model/store_checks.dart';
 import '../model/test_store.dart';
 import '../test_navigation.dart';
-import 'message_list_checks.dart';
-import 'page_checks.dart';
+import 'checks.dart';
 import 'test_app.dart';
 
 void main () {
