--- test/widgets/message_list_checks.dart
@@ -1,8 +1,0 @@
-import 'package:checks/checks.dart';
-import 'package:zulip/model/narrow.dart';
-import 'package:zulip/widgets/message_list.dart';
-
-extension MessageListPageChecks on Subject<MessageListPage> {
-  Subject<Narrow> get initNarrow => has((x) => x.initNarrow, 'initNarrow');
-  Subject<int?> get initAnchorMessageId => has((x) => x.initAnchorMessageId, 'initAnchorMessageId');
-}
