--- test/widgets/profile_page_checks.dart
@@ -1,6 +1,0 @@
-import 'package:checks/checks.dart';
-import 'package:zulip/widgets/profile.dart';
-
-extension ProfilePageChecks on Subject<ProfilePage> {
-  Subject<int> get userId => has((x) => x.userId, 'userId');
-}
