+++ test/widgets/profile_test.dart
@@ -27,9 +27,7 @@ import '../model/test_store.dart';
 import '../stdlib_checks.dart';
 import '../test_images.dart';
 import '../test_navigation.dart';
-import 'message_list_checks.dart';
-import 'page_checks.dart';
-import 'profile_page_checks.dart';
+import 'checks.dart';
 import 'test_app.dart';
 
 late PerAccountStore store;
