--- test/widgets/store_checks.dart
@@ -1,8 +1,0 @@
-import 'package:checks/checks.dart';
-import 'package:flutter/widgets.dart';
-import 'package:zulip/widgets/store.dart';
-
-extension PerAccountStoreWidgetChecks on Subject<PerAccountStoreWidget> {
-  Subject<int> get accountId => has((x) => x.accountId, 'accountId');
-  Subject<Widget> get child => has((x) => x.child, 'child');
-}
